Move `MatchingTransaction` to `ethereum/mod.rs`
Since `MatchingTransaction` is implemented for all Ethereum blockchain
connectors that implement `LatestBlock`, `BlockByHash` and
`ReceiptByHash`, not just `Web3Connector`.
